+++ test/conflux/conflux.test.js
@@ -6,9 +6,7 @@ const { Conflux } = require('../../src');
 test('constructor()', async () => {
   const cfx = new Conflux();
 
-  expect(cfx.defaultEpoch).toEqual('latest_state');
   expect(cfx.defaultGasPrice).toEqual(undefined);
-  expect(cfx.defaultGas).toEqual(undefined);
   expect(cfx.provider.constructor.name).toEqual('BaseProvider');
 
   await expect(cfx.provider.call()).rejects.toThrow('call not implement');
@@ -17,14 +15,10 @@ test('constructor()', async () => {
 test('constructor({...})', () => {
   const cfx = new Conflux({
     url: 'http://localhost:12537',
-    defaultEpoch: 'latest_mined',
     defaultGasPrice: 100,
-    defaultGas: 1000000,
   });
 
-  expect(cfx.defaultEpoch).toEqual('latest_mined');
   expect(cfx.defaultGasPrice).toEqual(100);
-  expect(cfx.defaultGas).toEqual(1000000);
   expect(cfx.provider.constructor.name).toEqual('HttpProvider');
 });